+++ schema/sqlite.sql
@@ -6,7 +6,9 @@ CREATE TABLE flysystem_path (
   visibility TEXT,
   size INTEGER DEFAULT NULL,
   is_compressed INTEGER DEFAULT 1,
-  update_ts TEXT NOT NULL DEFAULT 0,
+  expiry TEXT DEFAULT NULL,
+  meta TEXT DEFAULT NULL,
+  update_ts TEXT NOT NULL DEFAULT CURRENT_TIMESTAMP,
   PRIMARY KEY (path_id)
 );