+++ tests/IntegrationTestCase.php
@@ -53,6 +53,11 @@ abstract class IntegrationTestCase extends TestCase
             foreach ($createSql as $sql) {
                 static::$testDbAdapter->query($sql);
             }
+
+            if (static::$driver === 'mysql') {
+                $timezone = date_default_timezone_get();
+                static::$testDbAdapter->query("SET time_zone = '{$timezone}'");
+            }
         }
 
         return static::$testDbAdapter;
@@ -128,8 +133,14 @@ abstract class IntegrationTestCase extends TestCase
 
         parent::setUp();
 
-        static::getTestDbAdapter()->query('DELETE FROM flysystem_path');
-        static::getTestDbAdapter()->query('DELETE FROM flysystem_chunk');
+        $db = static::getTestDbAdapter();
+        if (static::getDbDriverName() === 'sqlite') {
+            $db->query('DELETE FROM flysystem_path');
+            $db->query('DELETE FROM flysystem_chunk');
+        } else {
+            $db->query('TRUNCATE flysystem_path');
+            $db->query('TRUNCATE flysystem_chunk');
+        }
     }
 
     final protected static function assertRowCount(int $expectedCount, string $tableName, string $message = ''): void